Move the Class to resources
